Move function to inject.rs
